pyproject.toml
[project]
name = "batch-mcp-testing"
version = "1.0.0"
description = "动态 MCP 工具部署和测试框架"
authors = [{name = "AI Assistant", email = "ai@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "agentscope>=0.1.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "asyncio-mqtt>=0.16.0",
    "openai>=1.0.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "supabase>=2.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.6.0",
    "isort>=5.12.0"
]

test = [
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "coverage>=7.3.0"
]

[project.scripts]
batch-mcp = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0", 
    "flake8>=6.0.0",
    "mypy>=1.6.0",
    "isort>=5.12.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "coverage>=7.3.0"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]
